minor organizing
