+++ src/server/game/Entities/Creature/Creature.h
@@ -362,23 +362,6 @@ struct CreatureAIInstance
 
 typedef std::unordered_map<uint32, CreatureAIInstance> CreatureAIInstanceContainer;
 
-// `creature_ai_instance_door` table
-
-enum DoorType
-{
-    DOOR_TYPE_ROOM          = 0,    // Door can open if encounter is not in progress
-    DOOR_TYPE_PASSAGE       = 1,    // Door can open if encounter is done
-    DOOR_TYPE_SPAWN_HOLE    = 2,    // Door can open if encounter is in progress, typically used for spawning places
-    MAX_DOOR_TYPES
-};
-
-struct DoorData
-{
-    uint32 entry, bossId;
-    DoorType type;
-    uint32 boundary;
-};
-
 enum BoundaryType
 {
     BOUNDARY_NONE = 0,
