+++ src/server/game/Maps/ZoneScript.h
@@ -40,10 +40,10 @@ class ZoneScript
 {
         uint8 m_type;
     public:
-        ZoneScript();
+        ZoneScript() : m_type(ZONE_TYPE_MAP) { }
         virtual ~ZoneScript() = default;
 
-        void SetType(uint8 type);
+        void SetType(uint8 type) { m_type = type; }
         virtual uint32 GetCreatureEntry(uint32 /*guidlow*/, CreatureData const* data);
         virtual uint32 GetGameObjectEntry(uint32 /*guidlow*/, uint32 entry) { return entry; }
 
@@ -77,8 +77,6 @@ class ZoneScript
         virtual void SetData(uint32 /*DataId*/, uint32 /*Value*/) {}
 
         virtual void ProcessEvent(WorldObject* /*obj*/, uint32 /*eventId*/) {}
-
-        InstanceScript* ToInstanceScript();
 };
 
 #endif
